from local

+++ egs/librispeech/ASR/lm2am_cif/conformer.py
@@ -393,7 +393,7 @@ class Conformer(Transformer):
             
             if self.cif:
                 am_output = am_output.transpose(1, 2).contiguous()
-                padding_mask_tensor = memory_key_padding_mask.long()
+                padding_mask_tensor = memory_key_padding_mask
                 print("am_output size = ", am_output.size())
                 print("padding_mask = ", padding_mask_tensor.size())
                 print("padding_mask1 = ", padding_mask_tensor)
